Improve postbuild path search
Before, it would find the first instance of "msvc" in the directory
path, but since I happen to be using this repository within another
"*msvc*" directory, it was not using the correct path within the
repository itself! This ensures that the last instance of "msvc" is used,
which will definitely be within the repository.